requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
cryptography==43.0.1
sqlalchemy==2.0.32
